requirements.txt
groq>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pandas>=2.0.0
rich>=13.0.0
asyncio